% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteH5AD.R
\name{WriteH5AD}
\alias{WriteH5AD}
\title{WriteH5AD}
\usage{
WriteH5AD(seurat_object, env_path, output_path, assay = "RNA")
}
\arguments{
\item{seurat_object}{The Seurat object required to be converted}

\item{env_path}{A path used to execute python environment.}

\item{output_path}{The path stored converted AnnData object}

\item{assay}{The Name of the assay to use for expression data. Default "RNA".}
}
\value{
No data needed to be return
}
\description{
Write Seurat object into h5ad AnnData object format
}
\details{
Warn: To keep the integrity and reproducibility, the `WriteH5AD` func will use counts layer as adata.X.
#' @examples
\dontrun{

  sce <- readRDS(system.file("extdata", "sce.rds", package = "Gauguin"))
  WriteH5AD(sce,"path/to/your/python/env,"sce2.h5ad")
}
}
