% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadH5AD.R
\name{ReadH5AD}
\alias{ReadH5AD}
\title{ReadH5AD}
\usage{
ReadH5AD(h5ad_path, env_path)
}
\arguments{
\item{h5ad_path}{A path which stores H5AD file.}

\item{env_path}{A path used to execute python environment.}
}
\value{
seurat_obj A Seurat Object
}
\description{
Read H5AD AnnData Object,and convert to Seurat Object
}
\examples{
\dontrun{

h5ad_file <- system.file("extdata", "sce.h5ad", package = "Gauguin")
ReadH5AD(h5ad_file, env_path = "path/to/your/python/env")
}
}
